renamed the files getDormer to useGetDormersByUID and getDor merPermits to useGetDormerPermits
